Always scroll activeRef into view on mount.
First off, great work with kbar, tried it yesterday and really loved it! 👏 
I've encountered this little annoying bug where sometimes, when you close the command bar and open it again, the `activeIndex` is the right one and the `activeRef` is there, but the virtual list is scrolled down to the middle and so, you can't see the active element. I thought a good fix was to make sure that the `activeRef` is visible whenever you mount the command bar.