Fix IME issue in keyboard event handler (#274)
* Fix IME issue in keyboard event handler

Previously, when an IME user (e.g. Chinese, Japanese user) was picking a word, the 'ArrowUp' or 'ArrowDown' keydown would make selected item in the search results changed, and when user hit 'Enter', it would not only pick the word, but also submit at the same time. This commit adds a check for event.isComposing to the event handler to prevent this issue.

* lint

---------

Co-authored-by: Tim <timchang.tcc@gmail.com>