Fix Next.js bug where Safari scrolls to bottom when closing kbar (#284)
It's possible for the previous active element to be the same as the
current activeElement ref when blurring focus to prevent Safari from
scrolling down when the kbar is closed. This results in re-focusing the
input element.

This commit additionally checks that we didn't just blur the input
before closing, before re-focusing it.