fix: allow matches on words far in the label
The current settings passed to fuse.js, make it unable to match something appearing very far in the label. In other words, if the label is made of 10 words and the query of the user aim to select on the 10th word, you'll never get the item.

With this PR I suggest to relax this constraint by asking fuse.js to ignore the location when performing its searches.

It should fixes the issue #337.